+++ widget_driver/example/lib/widgets/coffee_community/coffee_community_page_driver.dart
@@ -31,7 +31,6 @@ class CoffeeCommunityPageDriver extends WidgetDriver {
     });
   }
 
-  @TestDriverDefaultValue(false)
   bool get isLoggedIn => _authService.isLoggedIn;
 
   @override