+++ widget_driver/example/lib/widgets/coffee_community/coffee_library/coffee_library_page_driver.dart
@@ -26,10 +26,8 @@ class CoffeeLibraryPageDriver extends WidgetDriver {
     _getCoffees();
   }
 
-  @TestDriverDefaultValue(false)
   bool get isFetching => _isFetching;
 
-  @TestDriverDefaultValue(10)
   int get numberOfCoffees => _coffees.length;
 
   @TestDriverDefaultValue(TestCoffee.testCoffee)