+++ widget_driver/example/lib/widgets/coffee_community/not_logged_in/not_logged_in_page_driver.g.dart
@@ -12,10 +12,13 @@ part of 'not_logged_in_page_driver.dart';
 
 class _$TestNotLoggedInPageDriver extends TestDriver implements NotLoggedInPageDriver {
   @override
-  String get notLoggedInText => 'Not logged in';
+  String get notLoggedInText => ' ';
 
   @override
-  String get registerNewAccountButtonText => 'Register a new account';
+  String get registerNewAccountButtonText => ' ';
+
+  @override
+  void didUpdateBuildContext(BuildContext context) {}
 
   @override
   void registerNewAccountTapped(BuildContext context) {}