+++ widget_driver/example/lib/widgets/my_app_driver.dart
@@ -15,6 +15,5 @@ class MyAppDriver extends WidgetDriver {
     _localization = Resolver(context).get(() => context.read<Localization>());
   }
 
-  @TestDriverDefaultValue('Coffee Demo App')
   String get appTitle => _localization.appTitle;
 }