+++ widget_driver/example/lib/widgets/playground/playground_test_widget/playground_test_child_widget_driver.dart
@@ -4,6 +4,5 @@ part 'playground_test_child_widget_driver.g.dart';
 
 @GenerateTestDriver()
 class PlaygroundTestChildWidgetDriver extends WidgetDriver {
-  @TestDriverDefaultValue('Text coming from TestDriver')
   String get theText => 'Text coming from RealDriver';
 }