+++ widget_driver/example/lib/widgets/playground/playground_test_widget/playground_test_child_widget_driver.g.dart
@@ -12,7 +12,7 @@ part of 'playground_test_child_widget_driver.dart';
 
 class _$TestPlaygroundTestChildWidgetDriver extends TestDriver implements PlaygroundTestChildWidgetDriver {
   @override
-  String get theText => 'Text coming from TestDriver';
+  String get theText => ' ';
 }
 
 class $PlaygroundTestChildWidgetDriverProvider extends WidgetDriverProvider<PlaygroundTestChildWidgetDriver> {