+++ widget_driver/example/lib/widgets/playground/playground_test_widget/playground_test_widget_driver.g.dart
@@ -12,7 +12,7 @@ part of 'playground_test_widget_driver.dart';
 
 class _$TestPlaygroundTestWidgetDriver extends TestDriver implements PlaygroundTestWidgetDriver {
   @override
-  String get buttonHeaderText => 'Button header text from TestDriver';
+  String get buttonHeaderText => ' ';
 }
 
 class $PlaygroundTestWidgetDriverProvider extends WidgetDriverProvider<PlaygroundTestWidgetDriver> {