+++ widget_driver_generator/lib/src/utils/field_element_extensions.dart
@@ -1,0 +1,9 @@
+import 'package:analyzer/dart/element/element.dart';
+// ignore: implementation_imports
+import 'package:analyzer/src/dart/element/element.dart';
+
+extension FieldElementExtension on FieldElement {
+  /// All getters are also treated as fields but have explicit getters.
+  bool get isRedundantToPropertyAccessorElement =>
+      setter is! PropertyAccessorElementImpl_ImplicitSetter && getter is! PropertyAccessorElementImpl_ImplicitGetter;
+}